+++ service/submitqueue/orchestrator/server/BUILD.bazel
@@ -50,12 +50,12 @@ go_library(
         "//submitqueue/extension/conflict/fake:go_default_library",
         "//submitqueue/extension/conflict/none:go_default_library",
         "//submitqueue/extension/conflict/pathoverlap:go_default_library",
-        "//submitqueue/extension/scorer:go_default_library",
-        "//submitqueue/extension/scorer/composite:go_default_library",
-        "//submitqueue/extension/scorer/fake:go_default_library",
-        "//submitqueue/extension/scorer/heuristic:go_default_library",
         "//submitqueue/extension/speculation/allocator/sticky:go_default_library",
         "//submitqueue/extension/speculation/generator/bestfirst:go_default_library",
+        "//submitqueue/extension/speculation/scorer:go_default_library",
+        "//submitqueue/extension/speculation/scorer/composite:go_default_library",
+        "//submitqueue/extension/speculation/scorer/fake:go_default_library",
+        "//submitqueue/extension/speculation/scorer/heuristic:go_default_library",
         "//submitqueue/extension/speculation/speculator:go_default_library",
         "//submitqueue/extension/speculation/speculator/standard:go_default_library",
         "//submitqueue/extension/storage:go_default_library",
@@ -121,7 +121,7 @@ go_test(
         "//submitqueue/extension/buildrunner:go_default_library",
         "//submitqueue/extension/changeprovider:go_default_library",
         "//submitqueue/extension/conflict:go_default_library",
-        "//submitqueue/extension/scorer:go_default_library",
+        "//submitqueue/extension/speculation/scorer:go_default_library",
         "//submitqueue/extension/speculation/speculator:go_default_library",
         "//submitqueue/extension/storage:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
