+++ service/submitqueue/orchestrator/server/profiles.go
@@ -45,12 +45,12 @@ import (
 	conflictfake "github.com/uber/submitqueue/submitqueue/extension/conflict/fake"
 	"github.com/uber/submitqueue/submitqueue/extension/conflict/none"
 	"github.com/uber/submitqueue/submitqueue/extension/conflict/pathoverlap"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer/composite"
-	scorerfake "github.com/uber/submitqueue/submitqueue/extension/scorer/fake"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer/heuristic"
 	"github.com/uber/submitqueue/submitqueue/extension/speculation/allocator/sticky"
 	"github.com/uber/submitqueue/submitqueue/extension/speculation/generator/bestfirst"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer/composite"
+	scorerfake "github.com/uber/submitqueue/submitqueue/extension/speculation/scorer/fake"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer/heuristic"
 	"github.com/uber/submitqueue/submitqueue/extension/speculation/speculator"
 	specstandard "github.com/uber/submitqueue/submitqueue/extension/speculation/speculator/standard"
 	"github.com/uber/submitqueue/submitqueue/extension/storage"
