+++ submitqueue/extension/speculation/generator/bestfirst/bestfirst.go
@@ -30,8 +30,8 @@ import (
 	"slices"
 
 	"github.com/uber/submitqueue/submitqueue/entity"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer"
 	"github.com/uber/submitqueue/submitqueue/extension/speculation/generator"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer"
 )
 
 // bestFirst generates candidate paths using independent dependency
