+++ submitqueue/extension/speculation/scorer/BUILD.bazel
@@ -3,7 +3,7 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["scorer.go"],
-    importpath = "github.com/uber/submitqueue/submitqueue/extension/scorer",
+    importpath = "github.com/uber/submitqueue/submitqueue/extension/speculation/scorer",
     visibility = ["//visibility:public"],
     deps = ["//submitqueue/entity:go_default_library"],
 )
