+++ submitqueue/extension/speculation/scorer/composite/BUILD.bazel
@@ -3,12 +3,12 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["scorer.go"],
-    importpath = "github.com/uber/submitqueue/submitqueue/extension/scorer/composite",
+    importpath = "github.com/uber/submitqueue/submitqueue/extension/speculation/scorer/composite",
     visibility = ["//visibility:public"],
     deps = [
         "//platform/metrics:go_default_library",
         "//submitqueue/entity:go_default_library",
-        "//submitqueue/extension/scorer:go_default_library",
+        "//submitqueue/extension/speculation/scorer:go_default_library",
         "@com_github_uber_go_tally//:go_default_library",
     ],
 )
@@ -19,7 +19,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//submitqueue/entity:go_default_library",
-        "//submitqueue/extension/scorer:go_default_library",
+        "//submitqueue/extension/speculation/scorer:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
         "@com_github_uber_go_tally//:go_default_library",
