+++ submitqueue/extension/speculation/scorer/composite/scorer.go
@@ -20,7 +20,7 @@ import (
 	"github.com/uber-go/tally"
 	"github.com/uber/submitqueue/platform/metrics"
 	"github.com/uber/submitqueue/submitqueue/entity"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer"
 )
 
 // ReduceFunc combines named scores into a single score.
