+++ submitqueue/extension/speculation/scorer/fake/fake.go
@@ -30,7 +30,7 @@ import (
 	"github.com/uber/submitqueue/platform/fakemarker"
 	"github.com/uber/submitqueue/submitqueue/core/changeset"
 	"github.com/uber/submitqueue/submitqueue/entity"
-	"github.com/uber/submitqueue/submitqueue/extension/scorer"
+	"github.com/uber/submitqueue/submitqueue/extension/speculation/scorer"
 )
 
 // Recognized marker token. See the package doc for the convention.
