+++ submitqueue/extension/speculation/scorer/mock/BUILD.bazel
@@ -3,11 +3,11 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["scorer_mock.go"],
-    importpath = "github.com/uber/submitqueue/submitqueue/extension/scorer/mock",
+    importpath = "github.com/uber/submitqueue/submitqueue/extension/speculation/scorer/mock",
     visibility = ["//visibility:public"],
     deps = [
         "//submitqueue/entity:go_default_library",
-        "//submitqueue/extension/scorer:go_default_library",
+        "//submitqueue/extension/speculation/scorer:go_default_library",
         "@org_uber_go_mock//gomock:go_default_library",
     ],
 )
